+++ composer.json
@@ -494,7 +494,6 @@
         "pop-schema/migrate-custompostmedia-wp": "0.7.13",
         "pop-schema/migrate-custompostmeta": "0.7.13",
         "pop-schema/migrate-custompostmeta-wp": "0.7.13",
-        "pop-schema/migrate-customposts": "0.7.13",
         "pop-schema/migrate-events-wp-em": "0.7.13",
         "pop-schema/migrate-everythingelse": "0.7.13",
         "pop-schema/migrate-locations": "0.7.13",
