+++ layers/Schema/packages/customposts/composer.json
@@ -16,8 +16,7 @@
     "prefer-stable": true,
     "require": {
         "php": "^8.0",
-        "pop-schema/queriedobject": "^0.8",
-        "pop-schema/migrate-customposts": "^0.8"
+        "pop-schema/queriedobject": "^0.8"
     },
     "require-dev": {
         "phpstan/phpstan": "^0.12.76",
