--- layers/Schema/packages/migrate-customposts/README.md
@@ -1,43 +1,0 @@
-# (Temporary package) Migrate code to package: Custom Posts
-
-Code to be migrated, from legacy to modern PHP code ([read more here](https://github.com/leoloso/PoP#codebase-migration)). Target package description:  Adds support for custom posts
-
-## Install
-
-Via Composer
-
-``` bash
-composer require pop-schema/migrate-customposts
-```
-
-## Development
-
-The source code is hosted on the [PoP monorepo](https://github.com/leoloso/PoP), under [`Schema/packages/migrate-customposts`](https://github.com/leoloso/PoP/tree/master/layers/Schema/packages/migrate-customposts).
-
-## Security
-
-If you discover any security related issues, please email leo@getpop.org instead of using the issue tracker.
-
-## Credits
-
-- [Leonardo Losoviz][link-author]
-- [All Contributors][link-contributors]
-
-## License
-
-GNU General Public License v2 (or later). Please see [License File](LICENSE.md) for more information.
-
-[ico-version]: https://img.shields.io/packagist/v/pop-schema/customposts.svg?style=flat-square
-[ico-license]: https://img.shields.io/badge/license-GPLv2-brightgreen.svg?style=flat-square
-[ico-travis]: https://img.shields.io/travis/pop-schema/customposts/master.svg?style=flat-square
-[ico-scrutinizer]: https://img.shields.io/scrutinizer/coverage/g/pop-schema/customposts.svg?style=flat-square
-[ico-code-quality]: https://img.shields.io/scrutinizer/g/pop-schema/customposts.svg?style=flat-square
-[ico-downloads]: https://img.shields.io/packagist/dt/pop-schema/customposts.svg?style=flat-square
-
-[link-packagist]: https://packagist.org/packages/pop-schema/customposts
-[link-travis]: https://travis-ci.org/pop-schema/customposts
-[link-scrutinizer]: https://scrutinizer-ci.com/g/pop-schema/customposts/code-structure
-[link-code-quality]: https://scrutinizer-ci.com/g/pop-schema/customposts
-[link-downloads]: https://packagist.org/packages/pop-schema/customposts
-[link-author]: https://github.com/leoloso
-[link-contributors]: ../../../../../../contributors
