--- layers/Schema/packages/migrate-customposts/composer.json
@@ -1,28 +1,0 @@
-{
-    "name": "pop-schema/migrate-customposts",
-    "type": "library",
-    "description": "(Temporary package) Migrate code to package https://github.com/pop-schema/customposts. Target package description: Adds support for custom posts",
-    "keywords": ["PoP", "customposts"],
-    "homepage": "https://github.com/PoPSchema/migrate-customposts",
-    "license": "GPL-2.0-or-later",
-    "authors": [
-        {
-            "name": "Leonardo Losoviz",
-            "email": "leo@getpop.org",
-            "homepage": "https://getpop.org"
-        }
-    ],
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "require": {
-        "php": "^8.0"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "0.8-dev"
-        }
-    },
-    "config": {
-        "sort-packages": true
-    }
-}
