--- layers/Schema/packages/migrate-customposts/migrate/initialization.php
@@ -1,13 +1,0 @@
-<?php
-namespace PoPSchema\CustomPosts;
-
-class Initialization
-{
-    public function initialize()
-    {
-        /**
-         * Load the Library
-         */
-        include_once 'library/load.php';
-    }
-}
