--- layers/Schema/packages/migrate-customposts/migrate/library/functions/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'pop-engine-hooks.php';
