--- layers/Schema/packages/migrate-customposts/migrate/library/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'functions/load.php';
