+++ webservers/graphql-api-for-wp/.lando.yml
@@ -153,8 +153,6 @@ services:
         - >-
           ../../layers/Schema/packages/migrate-custompostmeta-wp:/app/wordpress/wp-content/plugins/graphql-api/vendor/pop-schema/migrate-custompostmeta-wp
         - >-
-          ../../layers/Schema/packages/migrate-customposts:/app/wordpress/wp-content/plugins/graphql-api/vendor/pop-schema/migrate-customposts
-        - >-
           ../../layers/Schema/packages/migrate-media:/app/wordpress/wp-content/plugins/graphql-api/vendor/pop-schema/migrate-media
         - >-
           ../../layers/Schema/packages/migrate-media-wp:/app/wordpress/wp-content/plugins/graphql-api/vendor/pop-schema/migrate-media-wp
